requirements.txt
dash==2.14.1
dash-bootstrap-components==1.5.0
pandas==2.1.3
numpy==1.26.2
plotly==5.18.0
umap-learn==0.5.4
scikit-learn==1.3.2
flask-cors==4.0.0